Add feature flag and switch for edit track redesign
