--- packages/web/src/components/edit-track/EditTrackForm/EditTrackForm.tsx
@@ -1,12 +1,0 @@
-import { Flex, Text } from '@audius/harmony'
-
-const messages = {
-  edit: 'Edit Track'
-}
-export const EditTrackForm = () => {
-  return (
-    <Flex>
-      <Text variant='display'>{messages.edit}</Text>
-    </Flex>
-  )
-}
